API_python/requirements.txt
flask==3.0.3
flask-cors==3.0.10
sqlalchemy==2.0.19
pyodbc==5.2.0  # Gebruik pyodbc als driver voor SQLAlchemy
azure-identity==1.12.0
azure-keyvault-secrets==4.9.0
requests==2.28.2
